+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -380,5 +380,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T11:12:15.108Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T11:18:13.725Z")
     r"""The date and time the resource was last updated"""