+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T11:12:15.108Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T11:18:13.725Z")
     r"""The date and time the resource was last updated"""