+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -92,5 +92,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T11:12:16.659Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T11:18:15.229Z")
     r"""The date and time the resource was created"""