+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQRGTRHAKGNH6XBM9B5BTXNT"
+        "01JQRH5PKEY5V31KG8NT8X971J"
     )
     r"""The id of the resource"""