+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQRGTRHBFWGJ9B5H1APFV8RB"
+        "01JQRH5PKFE97N2FFWZVB2K9MK"
     )
     r"""The id of the resource"""