+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQRGTRHCJ0H0AWDXZTGNZCJE"
+        "01JQRH5PKGWXXC5X93WTVNDNP6"
     )
     r"""The id of the resource"""